Strides have to be floats for all the types support

+++ include/mcm/tensor/tensor.hpp
@@ -285,7 +285,7 @@ namespace mv
         std::string toString() const override;
         virtual std::string getLogID() const override;
 
-        std::vector<unsigned> computeNumericStrides() const;
+        std::vector<float> computeNumericStrides() const;
         std::size_t computeTotalSize(unsigned int alignment = 16, bool base = false) const;
         std::size_t getClusterSize(unsigned int alignment = 16, bool base = false) const;
         void splitAcrossClusters(std::vector<Workload>, bool splitOverH, bool multicast);
